+++ src/Data/TypedEncoding/Instances/Support/Encode.hs
@@ -84,7 +84,8 @@ implEncodingEncodeEx' f = UnsafeMkEncoding Proxy $ implTranF f
         p = Proxy :: Proxy nm
 
 
-
+-- |
+-- @since 0.5.1.0
 implVerifyR :: (a -> Either err b) -> a -> Either err a
 implVerifyR fn a = 
      case fn a of